heaven/requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
brotlipy==0.7.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.3
cryptography==39.0.1
Django==4.1.7
django-cors-headers==3.14.0
django-redis==5.2.0
django-rest-framework==0.1.0
django-simpleui==2023.3.1
djangorestframework==3.14.0
dnspython==2.3.0
idna==3.4
libretranslatepy==2.1.1
lxml==4.9.2
mongoengine==0.27.0
mysqlclient==2.0.3
Pillow==9.4.0
pip==22.3.1
pycparser==2.21
pymongo==4.3.3
pyOpenSSL==23.0.0
PySocks==1.7.1
pytz==2022.7.1
redis==4.5.1
requests==2.28.1
setuptools==65.6.3
sqlparse==0.4.3
translate==3.6.1
urllib3==1.26.14
wheel==0.38.4
